Fix: Bug with lowercase int/Int

+++ _result/golang/Abteilung.go
@@ -1,6 +1,6 @@
 package golang
 
 type Abteilung struct {
-	AbteilungNr float64 `json:AbteilungNr`
-	Bezeichnung string  `json:Bezeichnung`
+	AbteilungNr int    `json:AbteilungNr`
+	Bezeichnung string `json:Bezeichnung`
 }
