+++ _result/golang/Buchung.go
@@ -1,10 +1,10 @@
 package golang
 
 type Buchung struct {
-	BuchungNr                      float64     `json:BuchungNr`
+	BuchungNr                      int         `json:BuchungNr`
 	Buchungsdatum                  string      `json:Buchungsdatum`
 	Belegdatum                     string      `json:Belegdatum`
-	Belegnummer                    float64     `json:Belegnummer`
+	Belegnummer                    int         `json:Belegnummer`
 	Kondition                      Kondition   `json:Kondition`
 	EinzahlungMitteilung           string      `json:EinzahlungMitteilung`
 	Verfalldatum                   string      `json:Verfalldatum`
@@ -12,19 +12,19 @@ type Buchung struct {
 	Zahlungsart                    Zahlungsart `json:Zahlungsart`
 	EsrKodierzeile                 string      `json:EsrKodierzeile`
 	EsrPruefziffer                 string      `json:EsrPruefziffer`
-	Herkunft                       float64     `json:Herkunft`
+	Herkunft                       int         `json:Herkunft`
 	Buchungen                      string      `json:Buchungen`
-	BuchungszeileNr                float64     `json:BuchungszeileNr`
+	BuchungszeileNr                int         `json:BuchungszeileNr`
 	Buchungsart                    Buchungsart `json:Buchungsart`
 	Auftrag                        Auftrag     `json:Auftrag`
 	Belegart                       Belegart    `json:Belegart`
-	EsrNummer                      float64     `json:EsrNummer`
+	EsrNummer                      int         `json:EsrNummer`
 	EinzahlungName                 string      `json:EinzahlungName`
 	HabenKonto                     Konto       `json:HabenKonto`
 	SollKonto                      Konto       `json:SollKonto`
 	Waehrung                       Waehrung    `json:Waehrung`
 	Kurs                           float64     `json:Kurs`
-	MahnCode                       float64     `json:MahnCode`
+	MahnCode                       int         `json:MahnCode`
 	MahnDatum                      string      `json:MahnDatum`
 	Steuercode                     Steuercode  `json:Steuercode`
 	SteuerbetragFW                 float64     `json:SteuerbetragFW`
