+++ _result/golang/ESRNummer.go
@@ -1,6 +1,6 @@
 package golang
 
 type ESRNummer struct {
-	ESRNr       float64 `json:ESRNr`
-	Bezeichnung string  `json:Bezeichnung`
+	ESRNr       int    `json:ESRNr`
+	Bezeichnung string `json:Bezeichnung`
 }
