+++ _result/golang/Liste.go
@@ -1,7 +1,7 @@
 package golang
 
 type Liste struct {
-	ListeNr     float64   `json:ListeNr`
+	ListeNr     int       `json:ListeNr`
 	Name        string    `json:Name`
 	Bezeichnung string    `json:Bezeichnung`
 	Art         string    `json:Art`
