+++ _result/golang/Rahmenvertrag.go
@@ -1,6 +1,6 @@
 package golang
 
 type Rahmenvertrag struct {
-	RahmenvertragNr float64 `json:RahmenvertragNr`
-	Bezeichnung     string  `json:Bezeichnung`
+	RahmenvertragNr int    `json:RahmenvertragNr`
+	Bezeichnung     string `json:Bezeichnung`
 }
