+++ _result/golang/Steuercode.go
@@ -1,7 +1,7 @@
 package golang
 
 type Steuercode struct {
-	SteuercodeNr float64 `json:SteuercodeNr`
+	SteuercodeNr int     `json:SteuercodeNr`
 	Bezeichnung  string  `json:Bezeichnung`
 	Prozent      float64 `json:Prozent`
 	InklusivMWST bool    `json:InklusivMWST`
