+++ _result/golang/Teilleistungstyp.go
@@ -1,6 +1,6 @@
 package golang
 
 type Teilleistungstyp struct {
-	TeilleistungstypNr float64 `json:TeilleistungstypNr`
-	Bezeichnung        string  `json:Bezeichnung`
+	TeilleistungstypNr int    `json:TeilleistungstypNr`
+	Bezeichnung        string `json:Bezeichnung`
 }
