+++ _result/golang/Verpackung.go
@@ -4,10 +4,10 @@ type Verpackung struct {
 	VerpackungNr string  `json:VerpackungNr`
 	Bezeichnung  string  `json:Bezeichnung`
 	Beschreibung string  `json:Beschreibung`
-	Laenge       float64 `json:Laenge`
-	Breite       float64 `json:Breite`
-	Hoehe        float64 `json:Hoehe`
-	Wandstaerke  float64 `json:Wandstaerke`
+	Laenge       int     `json:Laenge`
+	Breite       int     `json:Breite`
+	Hoehe        int     `json:Hoehe`
+	Wandstaerke  int     `json:Wandstaerke`
 	Gewicht      float64 `json:Gewicht`
 	Datei        string  `json:Datei`
 }
