+++ _result/golang/Verpackungsstufe.go
@@ -1,7 +1,7 @@
 package golang
 
 type Verpackungsstufe struct {
-	VerpackungsstufeNr float64 `json:VerpackungsstufeNr`
-	Bezeichnung        string  `json:Bezeichnung`
-	Indikator          float64 `json:Indikator`
+	VerpackungsstufeNr int    `json:VerpackungsstufeNr`
+	Bezeichnung        string `json:Bezeichnung`
+	Indikator          int    `json:Indikator`
 }
